added to .gitignor .DS_Store replaced sales types everywhere from "int" to "long"

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {46, 93, 169, 92, 154};
+        long[] sales = {46, 93, 169, 92, 154};
         SalesManager salesManager = new SalesManager(sales);
-        int res = salesManager.max();
+        long res = salesManager.max();
         System.out.println("Максимальное количество продаж: " + res);
     }
 }
